@@ -1,46 +1,47 @@
 [project]
 authors = [{ name = "Lorenzo Wormer", email = "l.wormer@mailbox.org" }]
-requires-python = ">=3.10,<3.12"
+requires-python = ">=3.12"
 dynamic = ["version"]
 dependencies = [
-    "pandas<3.0,>=2.2.0",
-    "PyYAML<7.0,>=6.0",
-    "openpyxl<4.0.0,>=3.0.9",
-    "country-converter<2.0,>=1.2",
-    "pydantic<3.0,>=2.5.3",
-    "streamlit<2.0.0,>=1.7.0",
-    "tqdm<5.0.0,>=4.62.3",
-    "sqlalchemy<3.0.0,>=2.0.7",
-    "strenum<1.0.0,>=0.4.10",
-    "stqdm<1.0.0,>=0.0.5",
-    "structlog<24.0.0,>=23.1.0",
-    "webcolors<2.0,>=1.13",
-    "deep-translator<2.0.0,>=1.11.0",
-    "babel<3.0.0,>=2.12.1",
-    "numpy<2.0.0,>=1.24.3",
-    "pdfkit<2.0.0,>=1.0.0",
-    "imgkit<2.0.0,>=1.2.3",
-    "typing-extensions<5.0.0,>=4.7.1",
-    "plotly<6.0.0,>=5.16.1",
-    "GitPython>=3.1.40",
-    "beautifulsoup4>=4.12.2",
-    "importlib-metadata>=6.11.0",
-    "pandas-stubs>=2.1.4.231227",
-    "packaging>=23.2",
-    "requests>=2.31.0",
-    "pydantic-core>=2.14.6",
-    "notebook>=7.0.6",
-    "ipywidgets>=8.1.1",
-    "ipython>=8.21.0",
-    "sphinx>=7.2.6",
-    "polars>=0.20.21",
-    "XlsxWriter>=3.2.0",
-    "cloudpathlib>=0.18.1",
-    "duckdb-engine>=0.11.5",
-    "bidict>=0.23.1",
-    "lxml>=5.2.2",
-    "beartype>=0.18.5",
-    "yarl>=1.9.4",
+  "pandas<3.0,>=2.2.0",
+  "PyYAML<7.0,>=6.0",
+  "openpyxl<4.0.0,>=3.0.9",
+  "country-converter<2.0,>=1.2",
+  "pydantic<3.0,>=2.5.3",
+  "streamlit<2.0.0,>=1.7.0",
+  "tqdm<5.0.0,>=4.62.3",
+  "sqlalchemy<3.0.0,>=2.0.7",
+  "strenum<1.0.0,>=0.4.10",
+  "stqdm<1.0.0,>=0.0.5",
+  "structlog<25.0.0,>=24.4.0",
+  "webcolors<2.0,>=1.13",
+  "deep-translator<2.0.0,>=1.11.0",
+  "babel<3.0.0,>=2.12.1",
+  "pdfkit<2.0.0,>=1.0.0",
+  "imgkit<2.0.0,>=1.2.3",
+  "typing-extensions<5.0.0,>=4.7.1",
+  "plotly<6.0.0,>=5.16.1",
+  "GitPython>=3.1.40",
+  "beautifulsoup4>=4.12.2",
+  "importlib-metadata>=6.11.0",
+  "pandas-stubs>=2.1.4.231227",
+  "packaging>=23.2",
+  "requests>=2.31.0",
+  "pydantic-core>=2.14.6",
+  "notebook>=7.0.6",
+  "ipywidgets>=8.1.1",
+  "ipython>=8.21.0",
+  "sphinx>=7.2.6",
+  "polars>=0.20.21",
+  "XlsxWriter>=3.2.0",
+  "cloudpathlib>=0.18.1",
+  "duckdb-engine>=0.11.5",
+  "duckdb>=1.0.0",
+  "bidict>=0.23.1",
+  "lxml>=5.2.2",
+  "beartype>=0.18.5",
+  "yarl>=1.9.4",
+  "numpy>=2.0.0",
 ]
 name = "py-research"
 description = "Collection of utilities for R&D coding in Python 🐍"
