@@ -3,19 +3,20 @@
 from collections.abc import Hashable, Iterable, Mapping, Sequence
 from dataclasses import asdict, dataclass, field
 from functools import cached_property, reduce
+from inspect import getmodule
 from secrets import token_hex
+from types import ModuleType
 from typing import (
+    TYPE_CHECKING,
     Any,
     ClassVar,
     Generic,
     Self,
-    TypeAlias,
     TypeVar,
     TypeVarTuple,
     cast,
     get_args,
     get_origin,
-    get_type_hints,
     overload,
 )
 
@@ -51,7 +52,7 @@ Rec2 = TypeVar("Rec2", bound="Record")
 Rec3 = TypeVar("Rec3", bound="Record")
 Rec_cov = TypeVar("Rec_cov", covariant=True, bound="Record")
 
-RecordValue: TypeAlias = Rec | Iterable[Rec] | Mapping[Any, Rec]
+type RecordValue[Rec: Record] = Rec | Iterable[Rec] | Mapping[Any, Rec]
 Recs = TypeVar("Recs", bound=RecordValue)
 
 Sch = TypeVar("Sch", bound="Schema")
@@ -65,23 +66,39 @@ def _extract_record_type(hint: Any) -> type["Record"]:
 
     if issubclass(origin, Record):
         return origin
-    if isinstance(origin, Mapping):
+    if issubclass(origin, Mapping):
         return get_args(hint)[1]
-    if isinstance(origin, Iterable):
+    if issubclass(origin, Iterable):
         return get_args(hint)[0]
 
     raise ValueError("Invalid record value.")
 
 
+@dataclass(frozen=True)
+class TypeRef(Generic[Val]):
+    """Reference to a type."""
+
+    hint: str | type[Val]
+    ctx: ModuleType | None = None
+
+    def resolve(self) -> type[Val]:
+        """Resolve the type reference."""
+        generic = (
+            self.hint
+            if isinstance(self.hint, type)
+            else cast(type[Val], eval(self.hint, vars(self.ctx)))
+        )
+        assert issubclass(get_origin(generic) or type, Prop)
+        return get_args(generic)[0]
+
+
 @dataclass(kw_only=True, eq=False)
-class Prop(Generic[Rec, Val]):
-    """Define property of a record."""
+class Prop(Generic[Val]):
+    """Reference a property of a record."""
 
     primary_key: bool = False
 
     _name: str | None = None
-    _record_type: type["Record"] | None = None
-    _value_type: type[Val] | None = None
     _value: Val | None = None
 
     @property
@@ -91,113 +108,40 @@ class Prop(Generic[Rec, Val]):
         return self._name
 
     @property
-    def record_type(self) -> type[Rec]:
-        """Record type."""
-        assert self._record_type is not None, "Record type not set."
-        return cast(type[Rec], self._record_type)
-
-    @property
-    def value_type(self) -> type[Val]:
-        """Property value type."""
-        return cast(
-            type[Val],
-            self._value_type or self.record_type._value_types.get(self.name) or Any,
-        )
-
-    @property
     def value(self) -> Val:
         """Property value."""
         assert self._value is not None, "Property value not set."
         return self._value
 
-    def __setname__(self, _, name: str) -> None:  # noqa: D105
+    def __set_name__(self, _, name: str) -> None:  # noqa: D105
         self._name = name
 
     @overload
     def __get__(self, instance: "Record", owner: type["Record"]) -> Val: ...
 
     @overload
-    def __get__(
-        self: "Attr[Any, Val]", instance: None, owner: type[Rec]
-    ) -> "Attr[Rec, Val]": ...
-
-    @overload
-    def __get__(
-        self: "Rel[Any, Rec2, Any]", instance: None, owner: type[Rec]
-    ) -> "Rel[Rec, Rec2, Rec2]": ...
-
-    @overload
-    def __get__(
-        self: "Rel[Any, Iterable[Rec2], Any]", instance: None, owner: type[Rec]
-    ) -> "Rel[Rec, Iterable[Rec2], Rec2]": ...
-
-    @overload
-    def __get__(
-        self: "Rel[Any, Sequence[Rec2], Any]", instance: None, owner: type[Rec]
-    ) -> "Rel[Rec, Sequence[Rec2], Rec2]": ...
-
-    @overload
-    def __get__(
-        self: "Rel[Any, Mapping[Idx, Rec2], Any]", instance: None, owner: type[Rec]
-    ) -> "Rel[Rec, Mapping[Idx, Rec2], Rec2]": ...
+    def __get__(self, instance: None, owner: type[Rec]) -> "PropRef[Rec, Val]": ...
 
     @overload
     def __get__(self, instance: object | None, owner: type | None) -> Self: ...
 
     def __get__(  # noqa: D105
-        self, instance: "object | None", owner: type["Record"] | None
-    ) -> "Val | Attr | type[Record] | Rel | Self":
-        if owner is not None:
-            self._record_type = owner
-
+        self, instance: "object | None", owner: type | type[Rec] | None
+    ) -> "Val | PropRef[Rec, Val] | Self":
         if instance is not None:
             return self.value
-
+        elif owner is not None and issubclass(owner, Record):
+            return PropRef(
+                primary_key=self.primary_key,
+                _name=self._name,
+                record_type=cast(type[Rec], owner),
+                value_type=owner._static_types[self.name],
+            )
         return self
 
     def __set__(self, instance: "Record", value: Val) -> None:
         """Set the value of the property."""
         self._value = value
-
-    @staticmethod
-    def get_tag(rec_type: type["Record"]) -> "Rel | None":
-        """Retrieve relation-tag of a record type, if any."""
-        try:
-            rel = getattr(rec_type, "_rel")
-            return rel if isinstance(rel, Rel) else None
-        except AttributeError:
-            return None
-
-    @cached_property
-    def parent(self) -> "Rel | None":
-        """Parent relation of this Rel."""
-        return self.get_tag(self.record_type)
-
-    @cached_property
-    def path(self) -> list["Rel"]:
-        """Path from base record type to this Rel."""
-        return [
-            *(self.parent.path if self.parent is not None else []),
-            *([self] if isinstance(self, Rel) else []),
-        ]
-
-    def __rrshift__(self, left: "type[Record] | Rel") -> Self:
-        """Append a prop to the relation path."""
-        current_root = self.path[0]
-        new_root = left if isinstance(left, Rel) else left.rel(current_root.record_type)
-
-        prefixed_rel = reduce(
-            lambda r1, r2: Rel(**asdict(r2), _record_type=r1.a), self.path, new_root
-        )
-
-        return cast(
-            Self,
-            (
-                prefixed_rel
-                if isinstance(self, Rel)
-                else getattr(prefixed_rel.a, self.name)
-            ),
-        )
 
     def __hash__(self) -> int:
         """Hash the Prop."""
@@ -205,10 +149,93 @@ class Prop(Generic[Rec, Val]):
 
 
 @dataclass(kw_only=True, eq=False)
-class Attr(Prop[Rec, Val], sqla.ColumnClause[Val]):
-    """Define attribute of a record."""
+class PropRef(Prop[Val], Generic[Rec_cov, Val]):
+    """Reference a property of a record."""
+
+    record_type: type[Rec_cov]
+    value_type: TypeRef[Val]
+
+    @staticmethod
+    def get_tag(rec_type: type["Record"]) -> "RelRef | None":
+        """Retrieve relation-tag of a record type, if any."""
+        try:
+            rel = getattr(rec_type, "_rel")
+            return rel if isinstance(rel, RelRef) else None
+        except AttributeError:
+            return None
+
+    @cached_property
+    def parent(self) -> "RelRef | None":
+        """Parent relation of this Rel."""
+        return self.get_tag(self.record_type)
+
+    @cached_property
+    def path(self) -> list["RelRef"]:
+        """Path from base record type to this Rel."""
+        return [
+            *(self.parent.path if self.parent is not None else []),
+            *([self] if isinstance(self, RelRef) else []),
+        ]
+
+    def __rrshift__(self, left: "type[Record] | RelRef") -> Self:
+        """Append a prop to the relation path."""
+        current_root = self.path[0]
+        new_root = (
+            left if isinstance(left, RelRef) else left.rel(current_root.record_type)
+        )
+
+        prefixed_rel = reduce(
+            lambda r1, r2: RelRef(**asdict(r2), record_type=r1.a),
+            self.path,
+            new_root,
+        )
+
+        return cast(
+            Self,
+            (
+                prefixed_rel
+                if isinstance(self, RelRef)
+                else getattr(prefixed_rel.a, self.name)
+            ),
+        )
+
+
+@dataclass(kw_only=True, eq=False)
+class Attr(Prop[Val]):
+    """Define an attribute of a record."""
 
     attr_name: str | None = None
+
+    @overload
+    def __get__(self, instance: "Record", owner: type["Record"]) -> Val: ...
+
+    @overload
+    def __get__(self, instance: None, owner: type[Rec]) -> "AttrRef[Rec, Val]": ...
+
+    @overload
+    def __get__(self, instance: object | None, owner: type | None) -> Self: ...
+
+    def __get__(  # noqa: D105
+        self, instance: "object | None", owner: type | type[Rec] | None
+    ) -> "Val | AttrRef[Rec, Val] | Self":
+        if instance is not None:
+            return self.value
+        elif owner is not None and issubclass(owner, Record):
+            return AttrRef(
+                primary_key=self.primary_key,
+                _name=self._name,
+                record_type=cast(type[Rec], owner),
+                value_type=owner._static_types[self.name],
+                attr_name=self.attr_name,
+            )
+        return self
+
+
+@dataclass(kw_only=True, eq=False)
+class AttrRef(
+    Attr[Val], PropRef[Rec_cov, Val], sqla.ColumnClause[Val], Generic[Rec_cov, Val]
+):
+    """Reference an attribute of a record."""
 
     def __post_init__(self) -> None:  # noqa: D105
         # Initialize fields required by SQLAlchemy superclass.
@@ -234,28 +261,94 @@ class Attr(Prop[Rec, Val], sqla.ColumnClause[Val]):
         return sqla_types.to_instance(self._value_type)  # type: ignore
 
 
-@dataclass
-class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
-    """Define relation to another record."""
+@dataclass(kw_only=True, eq=False)
+class Rel(Prop[Recs]):
+    """Define a relation to another record."""
 
-    via: (
-        Attr
-        | Iterable[Attr]
-        | dict[Attr, "Attr"]
-        | type["Record"]
-        | "Rel"
-        | tuple["Rel", "Rel"]
-        | None
-    ) = None
-    index_by: "Attr | None" = None
+    via: "AttrRef | Iterable[AttrRef] | dict[AttrRef, AttrRef] | type[Record] | RelRef | tuple[RelRef, RelRef] | None" = (  # noqa: E501
+        None
+    )
+    index_by: "AttrRef | None" = None
 
-    _target_type: type["Record"] | None = None
+    @overload
+    def __get__(self, instance: "Record", owner: type["Record"]) -> Recs: ...
 
-    @cached_property
+    @overload
+    def __get__(
+        self: "Rel[Rec2]", instance: None, owner: type[Rec]
+    ) -> "RelRef[Rec, Rec2, Rec2]": ...
+
+    @overload
+    def __get__(
+        self: "Rel[Iterable[Rec2]]", instance: None, owner: type[Rec]
+    ) -> "RelRef[Rec, Iterable[Rec2], Rec2]": ...
+
+    @overload
+    def __get__(
+        self: "Rel[Sequence[Rec2]]", instance: None, owner: type[Rec]
+    ) -> "RelRef[Rec, Sequence[Rec2], Rec2]": ...
+
+    @overload
+    def __get__(
+        self: "Rel[Mapping[Idx, Rec2]]",
+        instance: None,
+        owner: type[Rec],
+    ) -> "RelRef[Rec, Mapping[Idx, Rec2], Rec2]": ...
+
+    @overload
+    def __get__(self, instance: None, owner: type[Rec]) -> "RelRef[Rec, Recs, Any]": ...
+
+    @overload
+    def __get__(self, instance: object | None, owner: type | None) -> Self: ...
+
+    def __get__(  # noqa: D105 # type: ignore
+        self, instance: "object | None", owner: type | type[Rec] | None
+    ) -> "Recs | RelRef | Self":
+        if instance is not None:
+            return self.value
+        elif owner is not None and issubclass(owner, Record):
+            typeref = owner._static_types.get(self.name)
+            target = (
+                _extract_record_type(typeref.resolve()) if typeref is not None else None
+            )
+            return RelRef(
+                primary_key=self.primary_key,
+                _name=self._name,
+                record_type=owner,
+                value_type=typeref or TypeRef(Iterable[target or Any]),
+                via=self.via,
+                index_by=self.index_by,
+                _target_type=target,
+            )
+        return self
+
+
+@dataclass(kw_only=True, eq=False)
+class RelRef(Rel[Recs], PropRef[Rec_cov, Recs], Generic[Rec_cov, Recs, Rec2]):
+    """Reference a relation to another record."""
+
+    _target_type: type[Rec2] | None = None
+
+    @property
     def target_type(self) -> type["Record"]:
-        """Target record type."""
-        assert self._target_type is not None, "Target record type not set."
-        return self._target_type
+        """Dynamic target type of the relation."""
+        if self._target_type is not None:
+            return self._target_type
+
+        match self.via:
+            case dict():
+                return next(iter(self.via.values())).record_type
+            case type():
+                return self.via
+            case RelRef():
+                if issubclass(self.record_type, self.via.target_type):
+                    return self.via.record_type
+                else:
+                    return self.via.target_type
+            case tuple():
+                return self.via[1].target_type
+            case _:
+                raise ValueError("Invalid relation definition.")
 
     @cached_property
     def fk_record_type(self) -> type["Record"]:
@@ -263,48 +356,51 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
         match self.via:
             case type():
                 return self.via
-            case Rel():
+            case RelRef():
                 return self.via.record_type
             case tuple():
                 return self.via[0].record_type
-            case dict() | Attr() | Iterable() | None:
+            case dict() | AttrRef() | Iterable() | None:
                 return self.record_type
 
     @cached_property
-    def fk_map(self) -> bidict["Attr", "Attr"]:
+    def fk_map(self) -> bidict["AttrRef", "AttrRef"]:
         """Map source foreign keys to target attrs."""
         target = self.target_type
 
         match self.via:
-            case type() | Rel() | tuple():
+            case type() | RelRef() | tuple():
                 return bidict()
             case dict():
                 return bidict(
                     {
-                        Attr(
-                            _record_type=self.record_type,
+                        AttrRef(
                             _name=fk.name,
-                            _value_type=fk.value_type,
+                            record_type=self.record_type,
+                            value_type=fk.value_type,
                         ): pk
                         for fk, pk in self.via.items()
                     }
                 )
-            case Attr() | Iterable():
+            case AttrRef() | Iterable():
                 attrs = self.via if isinstance(self.via, Iterable) else [self.via]
                 source_attrs = [
-                    Attr(
-                        _record_type=self.record_type,
+                    AttrRef(
                         _name=attr.name,
-                        _value_type=attr.value_type,
+                        record_type=self.record_type,
+                        value_type=attr.value_type,
                     )
                     for attr in attrs
                 ]
-                target_attrs = target._indexes()
+                target_attrs = target._indexes
                 fk_map = dict(zip(source_attrs, target_attrs))
 
                 assert all(
                     issubclass(
-                        self.record_type._value_types[fk_attr.name], pk_attr.value_type
+                        self.record_type._static_props[
+                            fk_attr.name
+                        ].value_type.resolve(),
+                        pk_attr.value_type.resolve(),
                     )
                     for fk_attr, pk_attr in fk_map.items()
                     if pk_attr.value_type is not None
@@ -314,24 +410,26 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
             case None:
                 return bidict(
                     {
-                        Attr(
-                            _record_type=self.record_type,
+                        AttrRef(
                             _name=f"{self._name}.{target_attr.name}",
-                            _value_type=target_attr.value_type,
+                            record_type=self.record_type,
+                            value_type=target_attr.value_type,
                         ): target_attr
-                        for target_attr in target._indexes()
+                        for target_attr in target._indexes
                     }
                 )
 
     @cached_property
-    def inter_joins(self) -> dict[type["Record"], list[Mapping["Attr", "Attr"]]]:
+    def inter_joins(
+        self,
+    ) -> dict[type["Record"], list[Mapping["AttrRef", "AttrRef"]]]:
         """Intermediate joins required by this rel."""
         match self.via:
-            case Rel():
+            case RelRef():
                 # Relation is defined via other, backlinking relation
                 other_rel = self.via
                 assert isinstance(
-                    other_rel, Rel
+                    other_rel, RelRef
                 ), "Back-reference must be an explicit relation"
 
                 if issubclass(other_rel.target_type, self.record_type):
@@ -360,13 +458,13 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
                 # Relation is defined via relation table
                 back_rels = [
                     rel
-                    for rel in self.via._rels()
+                    for rel in self.via._rels
                     if issubclass(rel.target_type, self.record_type)
                     and len(rel.fk_map) > 0
                 ]
 
                 return {self.via: [back_rel.fk_map.inverse for back_rel in back_rels]}
-            case tuple() if has_type(self.via, tuple[Rel, Rel]):
+            case tuple() if has_type(self.via, tuple[RelRef, RelRef]):
                 # Relation is defined via back-rel + forward-rel
                 # on a relation table.
                 back, _ = self.via
@@ -377,10 +475,10 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
                 return {}
 
     @cached_property
-    def joins(self) -> list[Mapping["Attr", "Attr"]]:
+    def joins(self) -> list[Mapping["AttrRef", "AttrRef"]]:
         """Mappings of column keys to join the target on."""
         match self.via:
-            case Rel():
+            case RelRef():
                 # Relation is defined via other relation or relation table
                 other_rel = self.via
                 assert (
@@ -399,7 +497,7 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
                     # Relation is defined via all direct backlinks of given record type.
                     back_rels = [
                         rel
-                        for rel in self.via._rels()
+                        for rel in self.via._rels
                         if issubclass(rel.target_type, self.record_type)
                         and len(rel.fk_map) > 0
                     ]
@@ -409,7 +507,7 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
                 # Relation is defined via relation table
                 fwd_rels = [
                     rel
-                    for rel in self.via._rels()
+                    for rel in self.via._rels
                     if issubclass(rel.target_type, self.record_type)
                     and len(rel.fk_map) > 0
                 ]
@@ -419,7 +517,7 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
                 return [fwd_rel.fk_map for fwd_rel in fwd_rels]
 
             case tuple():
-                assert has_type(self.via, tuple[Rel, Rel])
+                assert has_type(self.via, tuple[RelRef, RelRef])
                 # Relation is defined via back-rel + forward-rel
                 # on a relation table.
                 _, fwd = self.via
@@ -446,14 +544,14 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
     def get_subdag(
         self,
         backlink_records: set[type["Record"]] | None = None,
-        _traversed: set["Rel"] | None = None,
-    ) -> set["Rel"]:
+        _traversed: set["RelRef"] | None = None,
+    ) -> set["RelRef"]:
         """Find all paths to the target record type."""
         backlink_records = backlink_records or set()
         _traversed = _traversed or set()
 
         # Get relations of the target type as next relations
-        next_rels = self.record_type._rels()
+        next_rels = self.record_type._rels
 
         for backlink_record in backlink_records:
             next_rels |= backlink_record._backrels_to_rels(self.record_type)
@@ -474,7 +572,7 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
             for rel in next_rel.get_subdag(backlink_records, _traversed)
         }
 
-    def __add__(self, other: "Rel[Any, Any, Rec3]") -> "RelTree[Rec2, Rec3]":
+    def __add__(self, other: "RelRef[Any, Any, Rec3]") -> "RelTree[Rec2, Rec3]":
         """Add another prop to the merge."""
         return RelTree({self}) + other
 
@@ -483,14 +581,14 @@ class Rel(Prop[Rec, Recs], Generic[Rec, Recs, Rec2]):
         return gen_int_hash(self)
 
 
-RelDict: TypeAlias = dict[Rel, "RelDict"]
+type RelDict = dict[RelRef, RelDict]
 
 
 @dataclass
 class RelTree(Generic[*RelTup]):
     """Tree of relations starting from the same root."""
 
-    rels: Iterable[Rel] = field(default_factory=set)
+    rels: Iterable[RelRef] = field(default_factory=set)
 
     def __post_init__(self) -> None:  # noqa: D105
         assert all(
@@ -517,61 +615,89 @@ class RelTree(Generic[*RelTup]):
 
         return tree
 
-    def __rrshift__(self, prefix: type["Record"] | Rel) -> Self:
+    def __rrshift__(self, prefix: type["Record"] | RelRef) -> Self:
         """Prefix all relations in the set with given relation."""
         rels = {prefix >> rel for rel in self.rels}
         return cast(Self, RelTree(rels))
 
-    def __add__(self, other: "Rel[Any, Any, Rec] | RelTree") -> "RelTree[*RelTup, Rec]":
+    def __add__(
+        self, other: "RelRef[Any, Any, Rec] | RelTree"
+    ) -> "RelTree[*RelTup, Rec]":
         """Append more rels to the set."""
         other = other if isinstance(other, RelTree) else RelTree([other])
         return RelTree([*self.rels, *other.rels])
 
     def __or__(
-        self: "RelTree[Rec]", other: "Rel[Any, Any, Rec2] | RelTree[Rec2]"
+        self: "RelTree[Rec]", other: "RelRef[Any, Any, Rec2] | RelTree[Rec2]"
     ) -> "RelTree[Rec | Rec2]":
         """Add more rels to the set."""
         other = other if isinstance(other, RelTree) else RelTree([other])
         return RelTree([*self.rels, *other.rels])
 
 
-class Record(Generic[Idx, Val]):
-    """Schema for a record in a database."""
+class RecordMeta(type):
+    """Metaclass for record types."""
 
-    _table_name: str
-    _type_map: dict[type, sqla.types.TypeEngine] = {
-        str: sqla.types.String().with_variant(
-            sqla.types.String(50), "oracle"
-        ),  # Avoid oracle error when VARCHAR has no size parameter
-    }
+    @property
+    def _id(cls) -> AttrRef:
+        """Default primary key attribute."""
+        return AttrRef(
+            _name="id",
+            record_type=cls,
+            value_type=TypeRef(int),
+            primary_key=True,
+        )
 
-    _value_types: dict[str, type]
-    _defined_attrs: set[Attr]
-    _defined_rels: set[Rel]
-
-    def __init_subclass__(cls) -> None:  # noqa: D105
-        # Retrieve property type definitions from class annotations
-        cls._value_types = {
-            name: get_args(hint)[0]
-            for name, hint in get_type_hints(cls)
-            if issubclass(get_origin(hint) or type, Attr)
+    @property
+    def _static_types(cls) -> dict[str, TypeRef]:
+        return {
+            name: TypeRef(hint, ctx=getmodule(cls))
+            for name, hint in cls.__annotations__.items()
+            if issubclass(get_origin(hint) or type, Prop)
+            or isinstance(hint, str)
+            and "Attr" in hint
+            or "Rel" in hint
         }
 
-        cls._defined_attrs = {
+    @property
+    def _static_props(cls) -> dict[str, PropRef]:
+        return {name: getattr(cls, name) for name in cls._static_types.keys()}
+
+    @property
+    def _dynamic_props(cls) -> dict[str, PropRef]:
+        return {
+            name: getattr(cls, name)
+            for name, prop in cls.__dict__.items()
+            if isinstance(prop, Prop)
+        }
+
+    @property
+    def _defined_props(cls) -> dict[str, PropRef]:
+        return {
+            name: prop
+            for name, prop in [
+                *cls._static_props.items(),
+                *cls._dynamic_props.items(),
+            ]
+        }
+
+    @property
+    def _defined_attrs(cls) -> set[AttrRef]:
+        return {
             getattr(cls, name)
-            for name, attr in cls.__dict__.items()
-            if isinstance(attr, Attr)
+            for name, attr in cls._defined_props.items()
+            if isinstance(attr, AttrRef)
         }
 
-        cls._defined_rels = {
+    @property
+    def _defined_rels(cls) -> set[RelRef]:
+        return {
             getattr(cls, name)
-            for name, rel in cls.__dict__.items()
-            if isinstance(rel, Rel)
+            for name, rel in cls._defined_props.items()
+            if isinstance(rel, RelRef)
         }
 
-        return super().__init_subclass__()
-
-    @classmethod
+    @property
     def _record_bases(cls) -> list[type["Record"]]:
         """Get all direct record superclasses of this class."""
         return [
@@ -580,37 +706,70 @@ class Record(Generic[Idx, Val]):
             if issubclass(c, Record) and c is not Record and c is not cls
         ]
 
-    @classmethod
-    def _base_pks(cls) -> dict[type["Record"], dict[Attr, Attr]]:
+    @property
+    def _base_pks(cls) -> dict[type["Record"], dict[AttrRef, AttrRef]]:
         return {
             base: {
-                Attr(
-                    _record_type=cls,
+                AttrRef(
                     attr_name=pk.name,
-                    _value_type=pk.value_type,
+                    record_type=cls,
+                    value_type=pk.value_type,
                     primary_key=True,
                 ): pk
-                for pk in base._indexes()
+                for pk in base._indexes
             }
-            for base in cls._record_bases()
+            for base in cls._record_bases
         }
 
-    @classmethod
-    def _rels(cls) -> set[Rel]:
+    @property
+    def _rels(cls) -> set[RelRef]:
         return reduce(
-            set.union, (c._rels() for c in cls._record_bases()), cls._defined_rels
+            set.union, (c._rels for c in cls._record_bases), cls._defined_rels
         )
 
-    @classmethod
-    def _attrs(cls) -> set[Attr]:
+    @property
+    def _attrs(cls) -> set[AttrRef]:
         return reduce(
-            set.union, (c._attrs() for c in cls._record_bases()), cls._defined_attrs
+            set.union, (c._attrs for c in cls._record_bases), cls._defined_attrs
         )
 
-    @classmethod
+    @property
     def _rel_types(cls) -> set[type["Record"]]:
         """Return all record types that are related to this record."""
-        return {rel.target_type for rel in cls._rels()}
+        return {rel.target_type for rel in cls._rels}
+
+    @property
+    def _indexes(cls) -> set[AttrRef]:
+        """Return the primary key attributes of this schema."""
+        defined_pks = {p for p in cls._attrs if p.primary_key}
+        return defined_pks or {cls._id}
+
+    if not TYPE_CHECKING:
+
+        def __getattr__(cls, name: str) -> AttrRef:
+            """Get and instantiate attribute by name."""
+            if name in cls.__dict__:
+                return cls.__dict__[name]
+
+            if name.startswith("__"):
+                return super().__getattribute__(name)
+
+            static_types = cls._static_types
+            if name in static_types:
+                return Prop(_name=name)
+
+            raise AttributeError(f"{cls.__name__} has no attribute '{name}'.")
+
+
+class Record(Generic[Idx, Val], metaclass=RecordMeta):
+    """Schema for a record in a database."""
+
+    _table_name: str
+    _type_map: dict[type, sqla.types.TypeEngine] = {
+        str: sqla.types.String().with_variant(
+            sqla.types.String(50), "oracle"
+        ),  # Avoid oracle error when VARCHAR has no size parameter
+    }
 
     @classmethod
     def _default_table_name(cls) -> str:
@@ -622,14 +781,9 @@ class Record(Generic[Idx, Val]):
         )
 
     @classmethod
-    def _indexes(cls) -> set[Attr]:
-        """Return the primary key attributes of this schema."""
-        return {p for p in cls._attrs() if p.primary_key}
-
-    @classmethod
     def _columns(cls, registry: orm.registry) -> list[sqla.Column]:
         """Columns of this record type's table."""
-        base_pk_cols = {pk for pks in cls._base_pks().values() for pk in pks.keys()}
+        base_pk_cols = {pk for pks in cls._base_pks.values() for pk in pks.keys()}
         table_attrs = (
             cls._defined_attrs
             | base_pk_cols
@@ -639,7 +793,11 @@ class Record(Generic[Idx, Val]):
         return [
             sqla.Column(
                 attr.name,
-                (registry._resolve_type(attr.value_type) if attr.value_type else None),
+                (
+                    registry._resolve_type(attr.value_type.resolve())
+                    if attr.value_type
+                    else None
+                ),
                 primary_key=attr.primary_key,
             )
             for attr in table_attrs
@@ -654,7 +812,7 @@ class Record(Generic[Idx, Val]):
             *(
                 sqla.ForeignKeyConstraint(
                     [attr.name for attr in rel.fk_map.keys()],
-                    [attr.name for attr in rel.target_type._indexes()],
+                    [attr.name for attr in rel.target_type._indexes],
                     table=(rel.target_type._table(metadata, subs)),
                     name=f"{cls._default_table_name()}_{rel._name}_fk",
                 )
@@ -667,7 +825,7 @@ class Record(Generic[Idx, Val]):
                     table=(base._table(metadata, subs)),
                     name=f"{cls._default_table_name()}_{base._default_table_name()}_inheritance_fk",
                 )
-                for base, pks in cls._base_pks().items()
+                for base, pks in cls._base_pks.items()
             ),
         ]
 
@@ -702,7 +860,7 @@ class Record(Generic[Idx, Val]):
 
         base_joins = [
             (base._joined_table(metadata, subs), pk_map)
-            for base, pk_map in cls._base_pks().items()
+            for base, pk_map in cls._base_pks.items()
         ]
         for target, pk_map in base_joins:
             table = table.join(
@@ -719,16 +877,18 @@ class Record(Generic[Idx, Val]):
         return table
 
     @classmethod
-    def _backrels_to_rels(cls, target: type[Rec]) -> set[Rel[Any, Any, Rec]]:
+    def _backrels_to_rels(
+        cls, target: type[Rec]
+    ) -> set[RelRef[Rec, Iterable[Self], Self]]:
         """Get all direct relations to a target record type."""
         return {
-            Rel(
-                _record_type=target,
-                _value_type=Iterable[cls],
-                _target_type=cls,
+            RelRef(
                 via=cls,
+                record_type=target,
+                value_type=TypeRef(Iterable[cls]),
+                _target_type=cls,
             )
-            for rel in cls._rels()
+            for rel in cls._rels
             if issubclass(target, rel.target_type)
         }
 
@@ -736,20 +896,23 @@ class Record(Generic[Idx, Val]):
     @classmethod
     def rel(
         cls, other: type[Rec], index: None = ...
-    ) -> Rel[Self, Iterable[Rec], Rec]: ...
+    ) -> RelRef[Self, Iterable[Rec], Rec]: ...
 
     @overload
     @classmethod
     def rel(
         cls, other: type[Rec], index: type[Key]
-    ) -> Rel[Self, Mapping[Key, Rec], Rec]: ...
+    ) -> RelRef[Self, Mapping[Key, Rec], Rec]: ...
 
     @classmethod
-    def rel(cls, other: type[Rec], index: type[Key] | None = None) -> Rel:
+    def rel(cls, other: type[Rec], index: type[Key] | None = None) -> RelRef:
         """Dynamically define a relation to another record type."""
         value_type = Mapping[index, other] if index is not None else Iterable[other]
-        return Rel(
-            _record_type=cls, _value_type=value_type, _target_type=other, via=other
+        return RelRef(
+            via=other,
+            record_type=cls,
+            value_type=TypeRef(value_type),
+            _target_type=other,
         )
 
     @classmethod
@@ -767,7 +930,7 @@ class Schema:
     def __init_subclass__(cls) -> None:  # noqa: D105
         subclasses = get_subclasses(cls, max_level=1)
         cls._record_types = {s for s in subclasses if isinstance(s, Record)}
-        cls._rel_record_types = {rr for r in cls._record_types for rr in r._rel_types()}
+        cls._rel_record_types = {rr for r in cls._record_types for rr in r._rel_types}
         super().__init_subclass__()
 
 
@@ -778,16 +941,19 @@ class Require:
     present: bool = True
 
 
-class DynRecordMeta(type):
+class DynRecordMeta(RecordMeta):
     """Metaclass for dynamically defined record types."""
 
-    def __getitem__(cls, name: str) -> Attr:
+    def __getitem__(cls, name: str) -> AttrRef:
         """Get dynamic attribute by dynamic name."""
-        return Attr(_record_type=cls, _name=name, _value_type=Any)
+        return AttrRef(_name=name, record_type=cls, value_type=TypeRef(Any))
 
-    def __getattribute__(cls, name: str) -> Attr:
+    def __getattr__(cls, name: str) -> AttrRef:
         """Get dynamic attribute by name."""
-        return Attr(_record_type=cls, _name=name, _value_type=Any)
+        if not TYPE_CHECKING and name.startswith("__"):
+            return super().__getattribute__(name)
+
+        return AttrRef(_name=name, record_type=cls, value_type=TypeRef(Any))
 
 
 class DynRecord(Record, metaclass=DynRecordMeta):  # noqa: N801
@@ -802,7 +968,7 @@ def dynamic_record_type(name: str, props: Iterable[Prop] = []) -> type[DynRecord
     return type(name, (DynRecord,), {p.name: p for p in props})
 
 
-AggMap: TypeAlias = dict[Attr[Rec, Any], Attr | sqla.Function]
+type AggMap[Rec: Record] = dict[AttrRef[Rec, Any], AttrRef | sqla.Function]
 
 
 @dataclass(frozen=True)
