@@ -1,59 +1,81 @@
 """Test db importing module."""
 
+from __future__ import annotations
+
 import json
 from datetime import date
-from typing import Any, Literal
+from typing import Literal
 
 import pytest
-
-from py_research.db import Attr, Record, Rel, RootMap
+from py_research.db import Attr, DataBase, Record, Rel, RootMap
+from py_research.db.importing import RelMap, SubMap, tree_to_db
 
 
 class SearchResult(Record):
     """Link search to a result."""
 
-    search: Rel[Any, "Search", "Search"]
-    result: Rel[Any, "Project", "Project"]
+    search: Rel[Search]
+    result: Rel[Project]
 
 
 class Search(Record[str, str]):
     """Defined search against the API."""
 
-    search_term: Attr[Any, str] = Attr(primary_key=True)
-    result_count: Attr[Any, int]
-    results: Rel[Any, list["Project"], "Project"] = Rel(via=SearchResult)
+    term: Attr[str] = Attr(primary_key=True)
+    result_count: Attr[int]
+    results: Rel[list[Project]] = Rel(via=SearchResult)
 
 
 class Task(Record):
     """Link search to a result."""
 
-    name: Attr[Any, str]
-    project: Rel[Any, "Project", "Project"]
-    assignee: Rel[Any, "User", "User"]
-    status: Attr[Any, Literal["todo", "done"]]
+    name: Attr[str]
+    project: Rel[Project]
+    assignee: Rel[User]
+    status: Attr[Literal["todo", "done"]]
 
 
 class User(Record):
     """A generic user."""
 
-    name: Attr[Any, str]
+    name: Attr[str]
+    age: Attr[int]
     tasks = Rel(via=Task.assignee)
+
+
+class Membership(Record):
+    """Link user to a project."""
+
+    member: Rel[User]
+    project: Rel[Project]
+    role: Attr[str]
 
 
 class Project(Record):
     """A generic project record."""
 
-    name: Attr[Any, str]
-    start: Attr[Any, date]
-    end: Attr[Any, date]
-    status: Attr[Any, Literal["planned", "started", "done"]]
+    name: Attr[str]
+    start: Attr[date]
+    end: Attr[date]
+    status: Attr[Literal["planned", "started", "done"]]
     tasks = Rel(via=Task.project)
+    members = Rel(via=Membership)
+    org: Rel[Organization]
+
+
+class Organization(Record):
+    """A generic organization record."""
+
+    name: Attr[str]
+    address: Attr[str]
+    city: Attr[str]
+    projects = Rel(via=Project.org)
 
 
 @pytest.fixture
 def nested_db_dict() -> dict:
     """Return nested data dict for import testing."""
-    with open("./nested_data.json") as f:
+    with open("tests/py_research/test_db/nested_data.json") as f:
         return json.load(f)
 
 
@@ -61,90 +83,65 @@ def nested_db_dict() -> dict:
 def root_table_mapping() -> RootMap:
     """Return root table mapping for import testing."""
     return RootMap(
-        table="searches",
-        id_type="hash",
-        id_attr="search_term",
+        rec=Search,
         push={
-            "resultCount": "result_count",
-            "search": "search_term",
-            "results": XMap(
-                table="projects",
-                id_type="uuid",
+            "resultCount": Search.result_count,
+            "search": Search.term,
+            "results": RelMap(
+                rel=Search.results,
                 push={
-                    "project_name": "name",
-                    "project_start": "start",
-                    "project_end": "end",
-                    "project_status": "status",
-                    "tasks": XMap(
-                        table="tasks",
+                    "project_name": Project.name,
+                    "project_start": Project.start,
+                    "project_end": Project.end,
+                    "project_status": Project.status,
+                    "tasks": RelMap(
+                        rel=Project.tasks,
                         push={
-                            "task_name": "name",
-                            "task_assignee": XMap(
-                                table="users",
-                                push="name",
-                                match_by_attrs="name",
+                            "task_name": Task.name,
+                            "task_assignee": RelMap(
+                                rel=Task.assignee,
+                                push=User.name,
+                                match=User.name,
                             ),
-                            "task_status": "status",
+                            "task_status": Task.status,
                         },
-                        id_type="uuid",
                     ),
-                    "members": XMap(
-                        table="users",
-                        push={"name", "age"},
-                        join_table_name="memberships",
-                        join_table_map={
-                            "role": "role",
+                    "members": RelMap(
+                        rel=Project.members,
+                        push={User.name, User.age},
+                        link=RootMap(
+                            rec=Membership,
+                            push={
+                                Membership.role,
+                            },
+                        ),
+                    ),
+                },
+                pull={
+                    Project.org: SubMap(
+                        push={
+                            "organization_name": Organization.name,
+                            "organization_address": Organization.address,
+                            "organization_city": Organization.city,
                         },
                     ),
                 },
-                ext_maps=[
-                    XMap(
-                        table="organizations",
-                        push={
-                            "organization_name": "name",
-                            "organization_address": "address",
-                            "organization_city": "city",
-                        },
-                        link_type="n-1",
-                        link_attr="organization",
-                        id_type="attr",
-                        id_attr="name",
-                    ),
-                ],
             ),
         },
     )
 
 
-def test_import_db_from_tree(nested_db_dict: dict, root_table_mapping: XMap):
+def test_import_db_from_tree(nested_db_dict: dict, root_table_mapping: RootMap):
     """Test importing nested data dict to database."""
-    db = tree_to_db(nested_db_dict, root_table_mapping, False)
-    assert isinstance(db, DB)
-    assert set(db.keys()) == {
-        "searches",
-        "projects",
-        "tasks",
-        "users",
-        "organizations",
-        "memberships",
-        "searches_projects",
-        "projects_tasks",
-    }
-    assert len(db["searches"].df()) == 1
-    assert len(db["projects"].df()) == 3
-    assert len(db["tasks"].df()) == 6
-    assert len(db["users"].df()) == 3
-    assert len(db["organizations"].df()) == 3
-    assert len(db["memberships"].df()) > 0
-    assert len(db["searches_projects"].df()) > 0
-    assert len(db["projects_tasks"].df()) > 0
-    assert db.relations == {
-        ("searches_projects", "results"): ("projects", "_id"),
-        ("searches_projects", "results_of"): ("searches", "_id"),
-        ("projects_tasks", "tasks"): ("tasks", "_id"),
-        ("projects_tasks", "tasks_of"): ("projects", "_id"),
-        ("memberships", "members"): ("users", "_id"),
-        ("memberships", "members_of"): ("projects", "_id"),
-        ("projects", "organization"): ("organizations", "_id"),
-    }
-    assert db.join_tables == {"searches_projects", "projects_tasks", "memberships"}
+    db, conflicts = tree_to_db(nested_db_dict, root_table_mapping)
+
+    assert isinstance(db, DataBase)
+    assert len(conflicts) == 0
+
+    assert len(db[Search]) == 1
+    assert len(db[Project]) == 3
+    assert len(db[Task]) == 6
+    assert len(db[User]) == 3
+    assert len(db[Organization]) == 3
+    assert len(db[Membership]) > 0
+    assert len(db[SearchResult]) > 0
